api-server/requirements.txt
fastapi[all]
uvicorn
fastapi-utils
httpx
mysql-connector-python